pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = ""
authors = [
    { name = "Vasilii Stakrotckii", email = "142693897+inferno681@users.noreply.github.com" },
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.17.0"
pydantic-settings = "^2.7.1"
pyyaml = "^6.0.2"
redis = "^5.2.1"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
wemake-python-styleguide = "^1.0.0"
mypy = "^1.14.1"
isort = "^5.13.2"
types-pyyaml = "^6.0.12.20241230"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
